CHEF-3216: Update versioning to match new scheme
The Windows MSI has always dropped the leading zero from the
Chef release version for its use, e.g. 0.10.10 was 10.10.

This patch uses Major.Minor.Micro-Build, e.g. 10.12.0-1
Major, minor and build are taken from the Chef release and
build is an incremented version for changes to the MSI itself.